refactor(temp_sensor): migrate to lib.rs + in-module tests (xiao_esp32s3_sense pattern)
- Replace crates/temp_sensor_core with src/utils/ modules inside the main crate
- Add src/lib.rs (pub mod utils) and src/utils/{mac_utils,payload,recalibration}.rs
- Each module file contains the pure function + #[cfg(test)] mod tests inline
- Gate ESP32-specific deps with target cfg so cargo test --lib runs on the host
- Update Cargo.toml: [[bin]] + [lib] entries; move ESP-IDF deps to target cfg section
- Update README: test command, software layout, module descriptions
