+++ devices/xiao_esp32s3_temp_sensor/Cargo.toml
@@ -6,6 +6,14 @@ edition = "2021"
 resolver = "2"
 rust-version = "1.77"
 
+[[bin]]
+name = "xiao_esp32s3_get_temperature"
+path = "src/main.rs"
+
+[lib]
+name = "xiao_esp32s3_get_temperature"
+path = "src/lib.rs"
+
 [profile.release]
 opt-level = "s"
 
@@ -17,17 +25,18 @@ opt-level = "z"
 default = []
 wifi = []
 
+# ホスト (テスト用) で必要な依存のみ
 [dependencies]
 log = "0.4"
 anyhow = "1.0"
+
+# ESP32 ターゲット専用の依存 (ホストテスト時はコンパイル対象外)
+[target.'cfg(any(target_arch = "riscv32", target_arch = "xtensa"))'.dependencies]
 simple_ds18b20_temp_sensor = { git = "https://github.com/junkei-okinawa/esp-temp-sensor", rev = "74c87a2ee5933c9bc3ac2c6d58c730d909a36356" }
 esp-idf-svc = "0.51.0"
+esp-idf-sys = "0.36"
 toml-cfg = "=0.2"
-temp-sensor-core = { path = "crates/temp_sensor_core" }
 
 [build-dependencies]
 embuild = { version = "0.33", features = ["espidf"] }
 toml-cfg = "=0.2"
-
-[target.'cfg(any(target_arch = "riscv32", target_arch = "xtensa"))'.dependencies]
-esp-idf-sys = "0.36"
