--- devices/xiao_esp32s3_temp_sensor/crates/temp_sensor_core/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "temp-sensor-core"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-name = "temp_sensor_core"
-path = "src/lib.rs"
