+++ devices/xiao_esp32s3_temp_sensor/src/lib.rs
@@ -1,0 +1,5 @@
+/// ハードウェア非依存ロジック
+///
+/// ## モジュール構成
+/// - `utils`: 純粋関数群（MAC 変換・再キャリブレーション判定・ペイロード生成）
+pub mod utils;
