+++ devices/xiao_esp32s3_temp_sensor/src/utils/mod.rs
@@ -1,0 +1,10 @@
+/// ユーティリティモジュール
+/// ハードウェア非依存の純粋関数を提供
+
+pub mod mac_utils;
+pub mod payload;
+pub mod recalibration;
+
+pub use mac_utils::parse_mac;
+pub use payload::format_hash_payload;
+pub use recalibration::needs_recalibration;
